pyproject.toml
[tool.poetry]
name = "db_core"
version = "0.1.0"
description = ""
authors = ["github profile dayvagrant"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
sqlalchemy = "^1.3.18"
psycopg2-binary = "^2.8.5"
pymongo = "3.10.1"
"infi.clickhouse_orm" = "^2.1.0"
cryptography = "^3.1.1"
tenacity = "4.12.0"
boto3 = "^1.15.16"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
black = "^19.10b0"
pylint = "^2.5.3"
rope = "^0.17.0"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
